requirements.txt
beautifulsoup4==4.4.1
Django==1.9
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
mysqlclient==1.3.7
Pillow==3.0.0
requests==2.9.1
wheel==0.24.0
